feat(ingest): standardize sql type mappings
